+++ wss.py
@@ -62,8 +62,8 @@ def main(args=None):
         print('monitoring result file (%s) parsing failed' % file_path)
         exit(1)
 
-    snapshot_time = (result.end_time - result.start_time) / result.nr_snapshots
-    nr_shots_in_aggr = max(round(args.work_time * 1000 / snapshot_time), 1)
+    snapshot_time = float(result.end_time - result.start_time) / result.nr_snapshots
+    nr_shots_in_aggr = int(max(round(args.work_time * 1000 / snapshot_time), 1))
     target_snapshots = result.target_snapshots
 
     if nr_shots_in_aggr > 1:
